+++ engines/ags/detection.h
@@ -94,8 +94,6 @@ public:
 	bool canPlayUnknownVariants() const override {
 		return true;
 	}
-
-	GUI::OptionsContainerWidget *buildEngineOptionsWidgetStatic(GUI::GuiObject *boss, const Common::String &name, const Common::String &target) const override;
 };
 
 #endif
