+++ engines/ags/metaengine.h
@@ -54,6 +54,8 @@ public:
 	 */
 	Common::String getSavegameFile(int saveGameIdx, const char *target = nullptr) const override;
 
+	GUI::OptionsContainerWidget *buildEngineOptionsWidgetDynamic(GUI::GuiObject *boss, const Common::String &name, const Common::String &target) const override;
+
 	/**
 	 * Determine whether the engine supports the specified MetaEngine feature.
 	 *
